Test CRUD refactoring and testing updates
