+++ azext_iot/tests/product/test_command_test_create_unit.py
@@ -112,7 +112,7 @@ class TestTestCreateUnit(unittest.TestCase):
             str(context.exception),
         )
 
-    @mock.patch("azext_iot.product.providers.aics._process_models_directory")
+    @mock.patch("azext_iot.product.command_tests._process_models_directory")
     @mock.patch("azext_iot.sdk.product.aicsapi.AICSAPI.create_device_test")
     def test_create_with_default_badge_type_doesnt_check_models(
         self, mock_service, mock_process_models
@@ -127,7 +127,7 @@ class TestTestCreateUnit(unittest.TestCase):
 
         mock_process_models.assert_not_called()
         mock_service.assert_called_with(
-            True,
+            provisioning=True,
             body={
                 "validationType": "Certification",
                 "productId": self.product_id,
@@ -140,7 +140,7 @@ class TestTestCreateUnit(unittest.TestCase):
             },
         )
 
-    @mock.patch("azext_iot.product.providers.aics._process_models_directory")
+    @mock.patch("azext_iot.product.command_tests._process_models_directory")
     @mock.patch("azext_iot.sdk.product.aicsapi.AICSAPI.create_device_test")
     def test_create_with_pnp_badge_type_checks_models(
         self, mock_service, mock_process_models
@@ -161,7 +161,7 @@ class TestTestCreateUnit(unittest.TestCase):
 
         mock_process_models.assert_called_with("models_folder")
         mock_service.assert_called_with(
-            True,
+            provisioning=True,
             body={
                 "validationType": "Certification",
                 "productId": self.product_id,
@@ -183,8 +183,8 @@ class TestTestCreateUnit(unittest.TestCase):
             },
         )
 
-    @mock.patch("azext_iot.product.providers.aics._read_certificate_from_file")
-    @mock.patch("azext_iot.product.providers.aics._process_models_directory")
+    @mock.patch("azext_iot.product.command_tests._read_certificate_from_file")
+    @mock.patch("azext_iot.product.command_tests._process_models_directory")
     @mock.patch("azext_iot.sdk.product.aicsapi.AICSAPI.create_device_test")
     def test_create_with_cert_auth_reads_cert_file(
         self, mock_service, mock_process_models, mock_read_certificate
@@ -208,7 +208,7 @@ class TestTestCreateUnit(unittest.TestCase):
         mock_read_certificate.assert_called_with("mycertificate.cer")
         mock_process_models.assert_called_with("models_folder")
         mock_service.assert_called_with(
-            True,
+            provisioning=True,
             body={
                 "validationType": "Certification",
                 "productId": self.product_id,
@@ -232,8 +232,8 @@ class TestTestCreateUnit(unittest.TestCase):
             },
         )
 
-    @mock.patch("azext_iot.product.providers.aics._read_certificate_from_file")
-    @mock.patch("azext_iot.product.providers.aics._process_models_directory")
+    @mock.patch("azext_iot.product.command_tests._read_certificate_from_file")
+    @mock.patch("azext_iot.product.command_tests._process_models_directory")
     @mock.patch("azext_iot.sdk.product.aicsapi.AICSAPI.create_device_test")
     def test_create_with_tpm(
         self, mock_service, mock_process_models, mock_read_certificate
@@ -257,7 +257,7 @@ class TestTestCreateUnit(unittest.TestCase):
         mock_read_certificate.assert_not_called()
         mock_process_models.assert_called_with("models_folder")
         mock_service.assert_called_with(
-            True,
+            provisioning=True,
             body={
                 "validationType": "Certification",
                 "productId": self.product_id,
@@ -280,11 +280,10 @@ class TestTestCreateUnit(unittest.TestCase):
         )
 
     @mock.patch("azext_iot.sdk.product.aicsapi.AICSAPI.create_device_test")
-    @mock.patch("azext_iot.product.providers.aics._create_from_file")
+    @mock.patch("azext_iot.product.command_tests._create_from_file")
     def test_create_with_configuration_file(self, mock_from_file, mock_sdk_create):
         mock_file_data = {"mock": "data"}
         mock_from_file.return_value = mock_file_data
         create(self, configuration_file="somefile")
         mock_from_file.assert_called_with("somefile")
-        mock_sdk_create.assert_called_with(True, body=mock_file_data)
-
+        mock_sdk_create.assert_called_with(provisioning=True, body=mock_file_data)
