+++ azext_iot/tests/product/test_device_test_tasks_int.py
@@ -12,7 +12,7 @@ class TestProductDeviceTestTasks(LiveScenarioTest):
     def __init__(self, _):
         super(TestProductDeviceTestTasks, self).__init__(_)
         self.kwargs.update(
-            {"device_test_id": "4afc5755-3167-4c72-84f5-14b136e376e0"}
+            {"device_test_id": "9825514b-27e6-4563-8cd9-2b384f3ac10d"}
         )
 
     def setup(self):
